[IMP] odoopy: python2shell - select with decorator
The process that was used to "export" the python method into the shell
was error prone and not very clean.

With this commit, only the methods that are explicitly marked as
available will be "exported".

Importing methods from external libraries should be much easier and
safer.

There is also a new safety that will stop unautorized method to be
called in the eval.