[IMP] bring_back_masterbeta_to_master: use remote branch
The local master does not always have the latest changes of the remote
master branch (since I rarelly use it). After this commit the
master-beta branch will be built based on the current state of the master
branch on GitHub.

This commit also make zsh not ask for confirmation when deleting files
with the star notation (for some reason it does not respect the -f flag)