[IMP] oes: venv
reintroduce automatic venv usage, but only for the start and restore
commands.

Fetch has a bunch of possible paramters that are invalid for start.

It could be reintroduce (hence why I have not completly removed the code)
but that would imply adding more magic to the optionnal parameters
managment (not worth it for now)