["typo"] ewq == eza
I am szitchin to a qwerty layout
the eza command (Edit Zsh Alias) is not really practical to type
quickly, but the mov;enet is so ingraned in my figgers that there is no
way i'll be able to change it quickly.

I'll thus add this 'ewq' alias, as the fast way to enter 'eza'

This change is going to be such a pain in the ass at the start.