[IMP] go_prune_all: parallel
execute in parallel all the operations that can be made parallel
safelty. That is, operations executed on seperated repository  (note the
worktree prune not being done in parallel of the git_prune_branches on
the mulitverse repos)

this will give ugly logs, but it will also be faster, so all in all, it
is worth it.

before this commit, I had in general up to 4 cores being used at
50 to 80% load during the peak of the operations in that sections of the code, now
it is 100% all cores.