[editors] Switch to NVIM embeded into vscode
For how cool NVIM is, it is not great for the way I go through code if I
am looking for somehting that I don't know.
In that situation, vscode remain superior.
But I don;t want to throw away the work done on nvim , and the vim
motion emulator extension of vscode is not complete enough for my taste.

Hence switching to the neovim embeded extension, that basically just
runs neovim inside vscode.

There will still be future changes for sure, but this is putting the
ground work in place to have vscode+nvim  and nvim  use the same nvim
config without too many explosions.