Trying noice without the messages
This enables consulting :reg, :marks, and other feature that display
content in a side window, to be consulted in a slow fashion  (otherwise,
they kind of just disapear super fast), while not being spammed by file
saved messsages that stay visible for way too long.

this also means that I am only using noice to move the command line to
the middle of my screen. That may not be very efficient...