[REM] universe setup for the versioned odoo repos
This was necessary due to the large size of the Odoo community git repo,
that was reaching 50GB on my machine.

Keeping two seperate copies was too much and thus I have to pivot to
single approach. The multiverse one was chosen because of it's overall
higher usefulness and flexibility.